agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Oruhkl",email = "osasinnook@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langchain-community (>=0.3.17,<0.4.0)",
    "langgraph (>=0.2.72,<0.3.0)",
    "langchain-anthropic (>=0.3.7,<0.4.0)",
    "tavily-python (>=0.5.1,<0.6.0)",
    "langgraph-checkpoint-sqlite (>=2.0.4,<3.0.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "load-dotenv (>=0.1.0,<0.2.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "websockets (>=10.0.0,<14.0.0)",
    "web3 (>=7.8.0,<8.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
